requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
coqpit-config==0.2.1
coqui-tts==0.27.1
coqui-tts-trainer==0.3.1
cycler==0.12.1
Cython==3.1.4
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
encodec==0.1.1
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
grpcio==1.75.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
monotonic-alignment-search==0.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
networkx==3.5
num2words==0.5.14
numba==0.62.0
numpy==2.3.3
ollama==0.5.4
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==1.0.0
srt==3.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.56.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1